+++ src/components/ListItems/ResultItem/index.js
@@ -1,0 +1,3 @@
+import ResultItem from './ResultItem';
+
+export default ResultItem;
